Diplom_3. Исправление замечаний
